requirements.txt
numpy==1.26.4
scipy>=1.11.0

opencv-contrib-python==4.8.1.78
pillow==10.2.0
scikit-image>=0.21.0
mediapipe==0.10.33

torch==2.1.2
torchvision==0.16.2
timm==0.9.16

transformers>=4.36.0

fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-multipart==0.0.6
aiofiles>=23.2.0
httpx==0.26.0
python-dotenv==1.0.0

requests==2.31.0
slowapi==0.1.9
python-magic-bin==0.4.14

ddgs>=0.1.0

groq>=1.1.2

pytesseract==0.3.10
librosa>=0.10.0
moviepy>=1.0.0